requirements.txt
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
autogen-agentchat==0.4.5
autogen-core==0.4.5
autogen-ext==0.4.5
autogenstudio==0.4.0.6
azure-ai-inference==1.0.0b8
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.13.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
email_validator==2.2.0
et_xmlfile==2.0.0
fast-depends==2.4.12
fastapi==0.115.8
fastapi-cli==0.0.7
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonref==1.1.0
loguru==0.7.3
lxml==5.3.0
Mako==1.3.8
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
numpy==1.26.4
ollama==0.4.7
openai==1.61.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
phidata==2.7.10
pillow==11.1.0
playwright==1.50.0
portalocker==2.10.1
prettytable==3.14.0
protobuf==5.29.3
psycopg==3.2.4
puremagic==1.28
pyautogen==0.7.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.3
termcolor==2.5.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
XlsxWriter==3.2.2
youtube-transcript-api==0.6.3
zipp==3.21.0
